Добавление класса и файлов World, переделка Chunk в World.
