+++ WorldClass.cpp
@@ -1,0 +1,66 @@
+#include "WorldClass.h"
+
+TextureHolder::TextureHolder() { }
+
+void TextureHolder::load(Textures::ID id, const std::string& filename)
+{
+	std::unique_ptr<sf::Texture> texture(new sf::Texture());
+	texture->loadFromFile(filename);
+
+	gTextureMap.insert(std::make_pair(id, std::move(texture)));
+}
+
+sf::Texture& TextureHolder::get(Textures::ID id)
+{
+	auto found = gTextureMap.find(id);
+	return *found->second;
+}
+
+const sf::Texture& TextureHolder::get(Textures::ID id) const
+{
+	auto found = gTextureMap.find(id);
+	return *found->second;
+}
+
+// in future make by singleton/think about better decision
+extern TextureHolder texture_holder;
+
+bool Block::passable()
+{
+	return isPassable;
+}
+
+Block::Block()
+{
+	sf::Texture& texture = texture_holder.get(Textures::BLOCKS);
+	block.setTexture(texture);
+}
+
+void Block::drawU(sf::RenderWindow& window)
+{
+	block.setPosition(coordinates);
+	window.draw(block);
+	//std::cout << coordinates.x << " " << coordinates.y << std::endl;
+}
+
+void Block::set_coordinates(const sf::Vector2f& coord)
+{
+	coordinates = coord;
+}
+
+sf::FloatRect Block::getGlobalBound()
+{
+	return block.getGlobalBounds();
+}
+
+World::World() {
+	for (int i = 0; i < WORLD_HEIGHT; ++i)
+		for (int j = 0; j < WORLD_WIDTH; ++j)
+			tilemap[i][j] = nullptr;
+}
+
+void World::set_block (int x, int y, int id) {
+	tilemap[x][y] = new Block;
+	tilemap[x][y]->set_coordinates(sf::Vector2f(y * 32.f, x * 32.f));
+}
+
